feat: Redirect to auth if not authenticated
Introduces an effect in the app component to automatically redirect unauthenticated users to the auth page. This improves the user experience by ensuring they are prompted to log in or sign up if they navigate to protected areas without authentication.

Additionally, the `signUp` function in `supabase.service.ts` is updated to provide default `username` and `full_name` metadata during the sign-up process. This addresses potential race conditions and database constraint errors by ensuring a profile is adequately initialized server-side via a trigger, removing the need for client-side profile creation post-signup.